+++ src/api/types/addTrackToLocalityResponse.ts
@@ -1,5 +1,5 @@
-import { Locality } from "@/types/locality";
-import { Track } from "@/types/track";
+import { Locality } from "@/types/locality/locality";
+import { Track } from "@/types/track/track";
 
 export interface AddTrackToLocalityResponse {
     locality: Locality;
