+++ src/components/misc/MovingText.tsx
