--- src/styles/index.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const defaultStyles = StyleSheet.create({
-    container: {
-        flex: 1
-    }
-});
-
-export const utilsStyles = StyleSheet.create({
-    slider: {
-        height: 7,
-        borderRadius: 16
-    }
-});
